use {} to create empty string
Before this patch we simply used "" to pass/return an empty string. E.g.
    string foo(const string& s) {
        ...
        return "";
    }
    ...
    foo("");

This calls the constructor
    string::string(const char* str)
In principle this constructor then needs to call strlen(str) to figure out the
length of the passed string literal (and then allocate a buffer and copy the
string content, though in case of an empty string this isn't needed). If we
instead use the default string constructor then the compiler immediately knows
it should construct an empty string.

I checked:
 - Older gcc compilers (gcc-4.x and below) do generate better code when using
   the default constructor.
 - Gcc-5 and above are able to figure out the length of the "" string literal
   at compile time and optimize it to exactly the same code as the default
   constructor.
 - Visual studio cannot figure this out yet.

What I did in this patch was search for "" and replace most occurences with
string{} or string_ref{}. Very often this can be abbreviated to just {}. E.g.
in the example at the top we can replace "" with {} (so equally long) and we
get better performace for some compilers.