Optimized OSD rendering of the reverse bar
While working on OSD stuff, I discovered that the 'reverse.int.bar' widget
(this is the blue/red colored part of the reverse bar; with scale_factor=2 this
is a rectangle of 500x12 pixels) was 'constantly' (10 times per second) being
re-created. Normally (in 'record' mode), this widget has constant dimensions
(and it's mostly even invisible), so there's no reason why the underlying
graphic (a gradient) is being re-rendered.

This inefficiency was caused by the reverse.tcl script: the update_reversebar
proc would set the 'relw' property of 'reverse.bar.int' to 0 and soon after the
same property gets set to the proper value. Fixed by not setting the value 0.

Though after this fix, the widget was still often being redrawn. This was
caused by floating point rounding errors: in my tests the 'relw' property
seemed to frequently toggle between the values '1.0' and '0.999999999998'.
Fixed by rounding this value to 4 decimal places in the Tcl script.

After this fix, even when the reverse bar is invisible, it still once per
second causes (internal) (re-)rendering. This is because the text in the bar is
changing once per second. There also appear new tick marks in the bar once per
second. We render these changes even if the widget is currently invisible. The
reason for this is that we (potentially) need the size of the (invisible)
widgets to be able to calculate the position/size of child widgets. A possible
optimization is to delay rendering of invisible widgets till some (visible)
child widget actually depends on the size of its (invisible) parent widget.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12497 7512d703-071c-0410-b5c2-d2bdd895b7d1