Optimized zoomSurface() function
This function is used (when using the 'SDL' renderer) to display a OSD
rectangle with a PNG image when that image has to be scaled (up or down). It
uses bilinear interpolation.

Before this patch, the RGBA components of each pixel were interpolated
separately. Now we use the highly optimized PixelOperations::lerp() function
which operates on the combined RGBA components.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12498 7512d703-071c-0410-b5c2-d2bdd895b7d1