Eliminate conversion on image scaling
This patch eliminates a conversion on scaling a (freshly loaded) PNG image. See
previous commit for a much more detailed description.

As a bonus, the newly created (scaled) surface retains information on the alpha
channel (whether the source had an alpha channel or not). Before this patch we
unconditionally added a fully opaque alpha channel for images without alpha
channel. This is not as bad as it sounds because at a later point we had a
check for such useless alpha channels (see hasConstantAlpha()), but now we can
avoid that check. (The check itself is still useful because some PNG images
really have an alpha channel with a constant value for each pixel, sometimes
even an opaque value for each pixel.)

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12500 7512d703-071c-0410-b5c2-d2bdd895b7d1