Optimized creation of gradient (SDL renderer)
When a OSD rectangle with a gradient is constructed we internally create a
SDL_Surface containing this gradient. When this widget actually needs to be
drawn we blit this surface to the display surface. So usually calculating
gradients is a relatively infrequent operation. Though we have scripts that
constantly change the dimensions of OSD rectangles with gradients (e.g. the
reversebar or vu_meter script). And when the dimensions change we need to
recalculate the gradient. So it makes sense to optimize this operation.

In a previous patch we optimized the bilinear interpolation in the
zoomSurface() function by calculating the 4 RGBA pixel components
simultaneously (instead of interpolating each R,G,B,A component separately). In
the gradient routine we also perform a bilinear interpolation between the 4
corner pixels. Though an important difference is that for the gradient we
always interpolate between the same 4 colors, while in the zoom routine we
frequently switch colors. This property allows, in case of the gradient, to
calculate the interpolation incrementally which is much more efficient.

So although the zoom and gradient routines are very different, the idea to work
on the 4 RGBA components in parallel can still be applied. And that's what this
patch does. See comments in the code for more details about the new algorithm.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12504 7512d703-071c-0410-b5c2-d2bdd895b7d1