Parse softwaredb.xml using a sax parser
When I start openmsx like this:
   openmsx exit.tcl
with the file exit.tcl only containing this line:
   after time 0.001 exit
I found that parsing softwaredb.xml took about 43% of the total time. However
in my setup I had a full copy of softwaredb.xml in both my openmsx user and
system directory. So on more typical setups it might be only half of that
time. But still it's quite a lot.

Before this patch, we first converted the softwaredb.xml file to our own
internal XMLElement structure (this is roughly speaking a simplified version of
a DOM representation; the conversion of xml-file -> XMLElement is done using a
sax parser). Later this XMLElement structure is converted to our internal
RomDatabase structure (which is basically a std::map indexed on sha1sum). The
XMLElement structure is thrown away.

This patch skips the whole XMLElement step. It directly converts the
softwaredb.xml file to the RomDatabase structure (using a sax parser). This
code is more complex than the previous version, but it's also roughly twice as
fast (in the test above the softwaredb.xml step now only takes about 23% of the
total time).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12511 7512d703-071c-0410-b5c2-d2bdd895b7d1