Removed XMLElement::getId()
This method returned the value of the attribute "id", or if the element
doesn't have such an attribute, it (recursively) looked at the parent element.

Only in error message printing in Rom::init() it sometimes happened that the
element didn't directly have an "id" attribute itself (and in that case the
direct parent had it). But in Rom::init() we can as well print the name of the
device in the error message, that may even be clearer. All other uses of
getId() are replace with getAttribute("id").

The motivation for this change is to eventually get rid of the 'parent' member
in XMLElement. See next commits.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12521 7512d703-071c-0410-b5c2-d2bdd895b7d1