Added DeviceConfig class
The two remaining users of XMLElement::getParent() are
 a) XMLElement::getFileContext()
 b) HardwareConfig::createDevices()

Case a) is very specific to our use of XMLElement to represent different parts
of a MSX machine configuration: the machine and the extensions can have
different associated 'FileContexts' (meaning they should look for ROM files in
different directories). So FileContext doesn't really belong in a generic
XML-representation class.

In case b) we search up in the XML tree for <primary> or <secondary> tags to
figure out in which (sub)slot the current MSXDevice should be registered.

This patch is a preparation to remove both a) and b). It adds a DeviceConfig
class which is a combination of HardwareConfig and XMLElement. HardwareConfig
represents a machine (without extensions) or an extension (so everything that
belongs to the same HardwareConfig has the same FileContext). XMLElement
represents the configuration of the device itself.

This patch is rather large, but it are mostly mechanical changes: replace the
XMLElement parameter is each MSXDevice (sub)class with a DeviceConfig
parameter.

In follow up patches we will move FileContext from XMLElement to HardwareConfig
and also extend DeviceConfig with information about primary/secondary slots.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12523 7512d703-071c-0410-b5c2-d2bdd895b7d1