Added primary/secondary info to DeviceConfig, removed XMLElement::parent
Removed the last user of XMLElement::getParent() (see previous commit messages
for more details).

Together with the previous commit, the size of a XMLElement object has shrunken
by 16 bytes (or 8 bytes on 32 bit systems). We have quite a lot of XMLElements
in memory so, this may matter (though in the past there were still many more
while parsing the RomDatabase).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12525 7512d703-071c-0410-b5c2-d2bdd895b7d1