Removed MSXDevice::motherboard member
Before this patch each MSXDevice stored a reference to MSXMotherBoard twice:
once directly as MSXDevice::motherBoard and once indirectly via
DeviceConfig::HardwareConfig::motherBoard. This patch removes the (redundant)
first reference.

Actually until a few commits ago (when DeviceConfig was introduced) it wasn't
yet possible (or not easy) to remove this redundant member because the
HardwareConfig was only filled in via init() (so after the constructor).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12526 7512d703-071c-0410-b5c2-d2bdd895b7d1