Reduced #include dependencies / cleanup 'getter-chains'
The code contained quite some chains of getXXX() calls. For example:
    config.getMotherBoard().getCommandController();
This code requires an #include "MSXMotherBoard.hh", and often this chain was
the only reason for that #include. This patch introduces various helper methods
so that the code above can be written like
    config.getCommandController();
and the #include can be left out.

This patch also splits the DeviceConfig.hh file in a .hh and a .cc part so that
most of the #includes from DeviceConfig.hh can be moved to the .cc part.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12528 7512d703-071c-0410-b5c2-d2bdd895b7d1