Initial version of combine-dmk
This is still a work-in-progress tool!! This tool combines the different .DAT
files generated by READ-DMK.COM into a single .DMK disk image. The main job of
the tool is to convert all tracks to a common length (on a real disk each track
can have a different length, but the DMK format only support a single track
length for the whole disk).

The READ-DMK.COM tool is not yet in SVN. but it will come one of the following
days.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12544 7512d703-071c-0410-b5c2-d2bdd895b7d1