Fixed regression introduced in revision 12436
The symptom was corrupt graphics in windows (but only with a VC++ build, not
with a mingw build) with these settings:
  scale_factor=2
  scale_algorithm=simple
  blur=0
while displaying a 512 pixel wide screen mode (screen6/7 or screen0/80).

The problem was located in the Scale_1on1<>::operator() method (this one is
indeed used in the scenario above). The bug triggered when on x86 the asm
routine was _not_ executed (so when the c++ fall-back routine was executed).
This triggered in the vc++ build because that one only has a SSE-implementation
using streaming stores but no MMX version without streaming stores.

Actually now that I know what the problem is, I also found a way to trigger the
problem in Linux: run the above scenario, but start openmsx with the -nommx
commandline option.

Thanks to Vampier/Goonies for bisecting the problem!

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12548 7512d703-071c-0410-b5c2-d2bdd895b7d1