Drastically speedup first invocation of tileviewer::showall
The slowness was caused by the creation of a huge number of child widgets (all
in the same parent widget). For fast drawing we keep the collection of child
widgets sorted on ascending z-coordinate. Each time a new widget was created we
would do a full re-sort of this collection, while it would be enough to simply
insert the new widget in the correct location. With this change, the first
invocation of showall went from 23 seconds to a little over 0.2 seconds (100x
faster!).

Also when the z-coordinate of an existing widget was changed we did a full
re-sort. But it's a lot faster to 'rotate' the collection back into sorted
order. (Though I don't know any current Tcl script where this change would
cause a measurable speedup).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12559 7512d703-071c-0410-b5c2-d2bdd895b7d1