Load trainer definitions in a lazy way
On openmsx startup all .tcl files found in the share/script directory are
automatically sourced. I added some code to profile how long it takes to source
each file (this code is disabled by default).

Before this patch, the file trainer.tcl took the longest. This patch changes
trainer.tcl so that the actual trainer definitions (in the file
trainerdefs.tclinclude) are only loaded when actually needed. This reduces the
source time for trainer.tcl from about 6000us to 250us).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12568 7512d703-071c-0410-b5c2-d2bdd895b7d1