Fixed 2nd drive detection on National machines
During the MSX boot sequence, the disk roms detects whether the machine has 1
or 2 disk drives connected. Before this patch, an emulated National_FS-5500F2
machine only detected 1 drive.

The problem was located in the 'track00' signal. On Philips machines (I verified
on a NMS_8280) this signal is never active when the drive motor is turned off.
However the drive-detection routine on a National machine also needs a meaningful
track00 signal when the drive motors are turned off. (The write-protected signal
behaves in the same way on Philips machines).

We also tested the behavior of this signal on a Sony_HB-F500P machine (thanks
Quibus). And there the status of the motor also doesn't matter. So it *seems*
only Philips machines are special for this aspect.

This patch splits the 'connectionstyle=Philips' into 'Philips' and 'Sony'. The
style 'Philips' is as before, 'Sony' doesn't have this special track00 stuff.
Also all other connectionstyles don't have the track00 stuff anymore. This
fixes the National drive detection.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12583 7512d703-071c-0410-b5c2-d2bdd895b7d1