Only write .filecache file when it was changed
When openMSX was exited, before this patch, we unconditionally wrote the
.filecache file. Now we only write this file when it will be changed (when we
added a new entry, or removed an invalid entry in memory).

In a profile run of 'openmsx exit.tcl' (with exit.tcl doing exit after 0.01s
emulated time), writing .filecache took about 6% of the time. So IMHO it was
worth optimizing.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12588 7512d703-071c-0410-b5c2-d2bdd895b7d1