[1/2] Optimized parsing of .filecache
In the previous patch we avoided writing the .filecache file when possible. In
the same profile run as described in the previous patch, reading the .filecache
file takes about 7%, and there's no easy way to skip this step. So this and the
next patch optimize the .filecache parsing routine.

The largest part of the time was spend in parsing the file modification dates.
This patch replaces the implementation of the Date::fromString() function with
a more efficient implementation. Though the new implementation is more lines of
code and less easy to understand.

Now the largest part of the Date::fromString() function is spend in mktime()
(before it was sscanf()). I guess repeated calls to mktime() perform quite some
redundant work (e.g. all the stuff related to timezones), but I couldn't find a
portable way to speed this up.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12589 7512d703-071c-0410-b5c2-d2bdd895b7d1