Avoid zero-initializing SDL_Surfaces in some cases
Sometimes (e.g. when loading a PNG image), we create a SDL_Surface and
immediately overwrite the pixel data. Creating a SDL_Surface was done using
SDL_CreateSurface(). This function will allocate an internal buffer for the
pixel data and also zero-initialize this buffer. But we immediately overwrite
the pixels, so zero-initialization was useless.

It's possible to avoid this overhead by using SDL_CreateSurfaceFrom(). This
function accepts a buffer as input parameter and uses that instead of
allocating and clearing (and eventually freeing) an internal buffer. Though a
disadvantage is that you now also have to manage that buffer yourself.

We already had our own wrapper class SDLSurfacePtr around SDL_Surface. This
patch extends that wrapper with (optional) buffer management. SDLSurfacePtr now
has an extra constructor to create an uninitialized SDL_Surface. As a bonus
this constructor uses exceptions to report errors instead of returning a NULL
pointer, so the caller code actually gets simpler.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12591 7512d703-071c-0410-b5c2-d2bdd895b7d1