Added string_ref class (not yet used)
This class is based on a proposal for std::string_ref, see comments in code
for more details.

The motivation for this class is optimizing code sequences like:

    void f(const std::string& str); // function that accepts a string parameter
    ...
    vector<char> buffer;  // some buffer (could also e.g. be a mmapped file)
    ...                   // fill buffer (e.g. read a file)
    char* p = ...;        // search something in the buffer
    f(string(p, 10));     // pass part of the buffer to f()

The last line in this code does:
 - Create a temporary std::string object. This allocates heap memory and makes
   a copy of the actual string data (plus some extra string management
   overhead, e.g. for gcc initialize a reference count).
 - Pass this string to the function f(). f() takes the string by reference to
   const, meaning it only wants to observe the string (not change it).
 - After f() returned the temporary string object is destroyed, so the heap
   memory gets deallocated.

But f() only observes the string. So instead of making a copy of the data and
passing that, we could directly pass a reference to the original data. This is
exactly what the string_ref class can do. string_ref has an interface that is
very close to a read-only std::string (so it's often a drop-in replacement for
const std::string). But unless std::string it doesn't own the string data it
points to. So it's useful for e.g. parameter passing, but e.g. it can't (in
general) be used to store a string as a class member.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12592 7512d703-071c-0410-b5c2-d2bdd895b7d1