[1/2] Speedup RomInfo::nameToRomType()
This is not a terribly expensive function, though it's called quite a few times
during parsing of softwaredb.xml. Optimizing it (in this and the next patch)
gave a measurable speed improvement.

Before this patch we first looked up the input string in an 'alias' map to
(possibly) translate it to the 'proper' mapper name. Next that name was looked
up in a name->romtype map (both lookups are case insensitive). This patch
merges both maps into one, almost doubling the lookup speed.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12595 7512d703-071c-0410-b5c2-d2bdd895b7d1