Use pre-calculated table in BlipBuffer
Before this patch the BlipBuffer class calculated the table on the first use of
the class (but KeyClick uses BlipBuffer, independent of the resampler setting,
so in practice the BlipBuffer class is always used). On a x86 CPU, calculating
this table takes an insignificant amount of CPU time. However on less powerful
devices and especially on devices which lack hardware support for floating
point calculations (e.g. Dingoo) this calculation might measurably contribute
to startup time. So it makes sense to optimize this.

TODO for now I've checked in the generated file 'BlipTable.ii'. Ideally it
should be generated during the build process. OTOH the content of this file
will only rarely change, and I did include some compile-time checks that the
table was generated with the same constants than what will be used by the rest
of the openMSX code. So updating the build process has low priority.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12606 7512d703-071c-0410-b5c2-d2bdd895b7d1