Don't use floating point for getTicksTill() -> FixedPoint
Before this patch we called getTicksTillDouble() and then converted this
(double) result into a FixedPoint value. This patch adds a method that directly
converts to FixedPoint.

The old way was easier to implement and to prove correct, e.g. doesn't the new
way introduce overflow in any of the intermediate results? (I believe not).

This method is not used that much (I estimate dozens of times per second, but
not hundreds of times). On x86 this change probably won't make any difference,
but on platforms without hardware floating point support it might.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12607 7512d703-071c-0410-b5c2-d2bdd895b7d1