Use string_ref instead of std::string in TclObject
TclObject can (among other things) represent a string. Before this patch, when
you queried the string representation of a TclObject, we returned that string
as a std::string object. This will make a copy of the actual string content.
This patch instead returns the string as a string_ref object. So instead of
making a copy of the string, we just refer to the already existing string data
inside the TclObject. Of course this means the string_ref only remains valid
for as long as the TclObject is not changed. Usually this is not a problem
and it's easy to still make a copy (a std::string) if needed.

So the return type of TclObject::getString() was changed from std::string to
string_ref. Though some of the callers of this method are not yet prepared for
this, so in many cases this patch does something like:
   myObject.getString().str()
So it immediately converts the string_ref into a std::string. (This is also
what happened implicitly before this patch, so it's not slower). In follow up
patches many of these extra str() calls will be removed again.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12613 7512d703-071c-0410-b5c2-d2bdd895b7d1