Construct MSXException objects using string_ref
Many of the MSXException (and subclasses) objects we use are constructed with a
string literal. Others are constructed with a std::string (and this string is
often constructed via StringOp::Builder). In the former case, to avoid
constructing a std::string object at each caller site we already overloaded the
MSXException constructor with a variant that takes a 'const char*' parameter,
(the other constructor takes a const std::string&). In this patch we merge both
constructors into one that takes a strinf_ref parameter. This offers the same
advantage with less code.

Also changed StringOp::Builder to return a string_ref instead of a std::string.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12614 7512d703-071c-0410-b5c2-d2bdd895b7d1