Use string_ref in some StringOp functions
The various split** functions in StringOp take an input string, chop it into 2
or more pieces and return those pieces. Before this patch, each such piece was
returned as a std::string (meaning it has a new heap allocated memory block
initialized with a copy of (part of) the input string). This patch returns the
pieces as string_ref objects (meaning we can just refer to parts of the
existing string, no heap allocation, no memcpy).

Similarly the trim** functions in StringOp take an input string and return part
of it. There we can also use string_ref.

The startsWith/endsWith functions only inspect the std::string parameters, now
they take string_ref parameters. That way string literals can be passed without
constructing/destructing a temporary std::string object. And it's often the
case that the 2nd parameter in these functions is a string literal.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12616 7512d703-071c-0410-b5c2-d2bdd895b7d1