[1/n] string_ref stuff: mostly command and setting registration
Various optimizations enabled by string_ref. This and the next patches are not
very specific to one change. I more or less starting changing and made a commit
when I felt I finished a loosely coherent block. But there probably will be
some overlap in the patches of this series.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12618 7512d703-071c-0410-b5c2-d2bdd895b7d1