[9/n] string_ref stuff: various, often related to std::string::substr()
Calling std::string::substr() creates a new string which is a substring of the
original. This involves heap allocation plus memory copy. Often it's possible
to instead convert the std::string object into a string_ref and call substr()
on the string_ref object (the latter two operations are very cheap).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12626 7512d703-071c-0410-b5c2-d2bdd895b7d1