[14/n] string_ref stuff: Pluggable::getDescription()
All pluggables except one returned a string-literal as description. The
exception was KeyJoystick, that one includes the instance number in the
text. Though the text is equally descriptive without that number. So now
they all return a string_ref.

For now this is the last commit in the string_ref series. Though I'm sure
there are more opportunities to replace std::string with string_ref (so more
patches may follow in the future).

Compared to revision 12609
 - the startup speed has improved by a little over 3%
 - heap memory usage is within a couple of kB unchanged
 - size of the openmsx executable has decreased by 48kB

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12631 7512d703-071c-0410-b5c2-d2bdd895b7d1