Fixed possible crash in scalers when using SSE2
For example in Simple2xScaler.cc:453 we allocate we VLA on the stack and later
pass that buffer to the drawScanline() routine. The latter routine is
implemented using SSE2 instructions, so it needs 16-byte aligned buffers.

By luck this went always fine (even though the buffer was only guaranteed to be
4-byte aligned). Though today we got a bug report that on Fedora 17 32-bit it
crashed. After some debugging we found that the buffer was indeed misaligned.

I'm *guessing* we never saw this problem before because most of us are running
a 64-bit OS and there stack frames are always 16-byte aligned. At first sight
the code is simple enough to not require any register spills (I didn't actually
verify this), so the buffer is allocated at the start of the stack frame and
thus will also be 16-byte aligned. 32-bit x86 also has 16-byte aligned stack
frames, but it has less registers, so probably there it had to spill and this
caused the buffer to not be aligned anymore.

Fixed by requiring a 16-byte aligned buffer (using the VLA_ALIGNED macro).

Thanks to 'JaSeZ' for reporting the crash and to 'mth' for finding the cause. I
only implemented his suggestion for the fix.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12636 7512d703-071c-0410-b5c2-d2bdd895b7d1