Fixed assert in ResampleLQ.cc:246
We got crash on dingoo when running games that use the YM2413 (e.g. it crashed
when running the intro demo of 'Psycho World'). I could reproduce a crash on PC
when I set frequency=48000 and resampler=fast (like on dingoo). Though we're
not 100% sure this is the same problem.

Internally in the ResampleLQ class we used a 16.16 fixed point number to
represent the ration between emu and host sample rate. This value was
calculated as a floating point value and then converted to the fixed point
closest to the floating point (so either rounded up or down). Though there was
a problem when this value was rounded up. In that case in certain circumstances
by the accumulation of these rounding errors it could happen we did an
out-of-bound access of one emu sample. If we instead always round down this
problem cannot occur.

Thus fixed by always rounding down. As a bonus the calculation of the ratio
doesn't use floating point calculations anymore.

In the ResampleBlip code there was a similar fixed point number that could
either be rounded up or down (and after this patch it's always rounded down).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12643 7512d703-071c-0410-b5c2-d2bdd895b7d1