Lazily initialize LUT in V9990CmdEngine
The V9990 supports reasonably complex logical operations in its command engine.
To greatly speedup the calculations we use a lookup table. Each sub-LUT is 64kB
in size (input: source and destination byte, output: new destination byte). But
we need several of these sub-LUTs because there are 16 possible logical
operations. We also need different versions of the table for the different
screen modes (2,4,8,16 bpp). Combined this gives a LUT of 4MB.

When I profiled the startup performace of the 'Boosted_MSX2_EN' machine (or any
configuration that uses the gfx9000). I found that over 30% of the time was
spend in calculating the above LUT. Ironically the vast mojority of that table
is never used, not even when intensively running V9990 applications. To
increase startup performance (and to reduce memory usage) the LUT is now lazily
initialized in chunks of 64kB.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12646 7512d703-071c-0410-b5c2-d2bdd895b7d1