Improved/simplified PixelOperations<Pixel>::blend<w1,w2>()
There are two main changes:

a) blend<1,2> used to be calculated as a long chain of average operations (e.g.
like we still calculate <3,5> but then with a longer chain length). I looked at
the generated code, and the (long) chain of average operations generates worse
code than approximating <1,2> as <85,171> (the chain of avg introduces rounding
errors, so it's also only an approximation). (Historically, we didn't have the
special code for sum-of-weights-is-power-of-two, so at the time this routine
was better than the generic version).

b) We already had several other cases that we approximated as weights that sum
to 256 (e.g. <2,5> -> <73,183>). Now I've implemented a mechanism to
automatically calculate this approximation. So this removes some of the special
cases we already added and it works for any other pair of weights we will use
in the future.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12655 7512d703-071c-0410-b5c2-d2bdd895b7d1