Use rapid XML parser in XMLLoader (instead of libxml)
This patch replaces the libxml parser in the XMLLoader code with 'rapid xml'.
We already use 'rapid xml' for parsing softwaredb.xml (and it makes a big
performance difference, see 3852e778).

This patch simplifies the source code, while speeding it up at the same time.
Though XMLLoader is not used (anymore) to load big XML files, so I don't expect
a noticeable total improvement. However the main motivation for this patch is
to eventually get rid of our dependency on libxml.

After this patch we still use libxml in two locations:
* XMLElement::XMLEscape():
   This will escape all character with special meaning. E.g. '<' is replaced
   with '&lt;'. It should be easy to write our own version of this routine.
* parsing in CliConnection class:
   The commands received over a CliComm connection are actually formatted as an
   XML file. Though this data should already be parsed before the whole XML
   document is received. Libxml can do this, but rapidxml cannot. It also
   shouldn't be too difficult to replace this parser with our own (we only need
   to recognize a small subset of XML features), but for the moment this is low
   priority for me.