Removed redundant (leftover) code in SHA1::finalize()
Revision 12587 changed the sha1 code to not represent the result of a sha1
calculation as a string, but keep it as a 160-bit number (and only convert to
string when really needed). Everything was correct, but I forgot to remove a
few lines of code, so between rev12587 and this patch we still did a redundant
sha1->string conversion (the result of this conversion was not used).

In this patch I also made a tiny optimization in the (non-redundant)
Sha1Sum::toString() method.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12656 7512d703-071c-0410-b5c2-d2bdd895b7d1