Use string_ref in Layer::getName() and Display::findLayer()
The next patch will use Display::findLayer() more, so in this patch we first
optimize it a bit. (No need to construct/destruct temporary std::string
objects).

The method Layer::getName() is renamed to getLayerName(). The reason is that in
the next patch we have a class that inherits from both MSXDevice and Layer.
Without the rename, both base classes would have a getName() method and it's
not so easy to override only one of the two in the subclass. See this page for
some background info:
  http://stackoverflow.com/questions/2004820/inherit-interfaces-which-share-a-method-
(There are alternative solutions, but a rename is simpler.)

This patch simplifies the layer names a bit. And also for the next patch it's
useful that the names of the 'GLSnow' and 'SDLSnow' layers are the same. So
both are renamed to 'snow'.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12658 7512d703-071c-0410-b5c2-d2bdd895b7d1