Initial implementation of Video9000
When the new 'video9000' extension is inserted, you get a new videosource:
'Video9000' (in addition to 'MSX' and 'GFX9000'). This new videosource shows
the video output of the Video9000 cartridge. This cartridge can show the
gfx9000 output, one of the 3 video inputs of the Video9000 cartridge, or even
the gfx9000 superimposed over one of the 3 inputs. Normally the MSX V99x8
output is connected to one of the 3 video9000 inputs.

Current limitations:
 - We only emulate one video input, and connect this to the V99x8. So in the
   current implementation it is as-if the V99x8 is connected to all 3 inputs
   simultaneous. (low priority to implement)
 - 'screenshot -raw' and video recording don't work yet (will be fixed in the
   next few patches)
 - Superimpose is not yet implemented (will take a bit longer to implement)

So currently, when you select the Video9000 videosource (and this is the new
default), you can switch between V99x8 and V9990 output by writing to IO port
0x6f.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12659 7512d703-071c-0410-b5c2-d2bdd895b7d1