Enable 'screenshot -raw' for video9000
As a bonus 'screenshot -raw' now also works when the 'laserdisc' videosource is
selected. (Note this is the laserdisc-only output, if you take a raw screenshot
of the V99x8 with superimposed laserdisc, you still only get the V99x8 output.
In a normal (non-raw) screenshot both laserdisc and V99x8 output are present).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12660 7512d703-071c-0410-b5c2-d2bdd895b7d1