Added video recording for video9000
This was actually easier than I thought. The hard work was already done in the
past when we allowed switching between V99x8 and V9990 video sources during
video recording. However on such a switch, the recorded video would 'flicker'
for a few frames (already before we added Video9000). That should be fixed now.

This patch also enables recording the LaserDisc videosource (this is laserdisc
only, not V99x8 superimposed over laserdisc).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12661 7512d703-071c-0410-b5c2-d2bdd895b7d1