experimental: Tcl syntax highlighting in the console
This is something I wanted to add for a very long time already. I already had
most of the building blocks (e.g. the TclParser class) lying around. But today
I finally integrated it in openMSX.

It's still experimental because there are quite some corner cases to get right.
Though I tested the individual building blocks already reasonably well. It may
also be the case that the color schema is not always well readable (depending
on the console background image and even the MSX image that shines through the
console background). It might be a good idea to make the color schema
configurable and even offer an option to disable the syntax highlighting
altogether.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12680 7512d703-071c-0410-b5c2-d2bdd895b7d1