Implemented tab-completion for nested Tcl commands
Some examples:
  if {[debug rea<tab>
  if {[debug read memory 0] == 0} {osd configu<tab>
  bind F6 {cyc<tab>
  poke 0x8123 7 ; poke 0x8124 [pee<tab>

Before this patch the examples above could not be completed because we only
split the line on ';' and tried to complete the last part (before the cursor).
Now we parse the Tcl command and start tab-completion at the last sub command.
So all the examples above should do the right thing now.

This patch also fixes a hang on invalid expressions like
  if {/       <-- missing operand before /
The algorithm that tried to complete incomplete expressions (so they can be
parsed) got in an infinite loop. And I also fixed a memory leak in the same
completion algorithm.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12690 7512d703-071c-0410-b5c2-d2bdd895b7d1