Extended PostProcessor to superimpose 2 VDP frames
We now have two different ways to superimpose
 - the old way use by laserdisc
 - the new way (will be used by video9000)

The main difference is that for laserdisc we want to keep the video frames
unchanged and superimpose a scaled MSX frame, while for video9000 we want to
superimpose the V99x8 and V9990 frames and scale the result. (An exception is
e.g. the 'simple' scaler, this one adds scanlines and it would look very weird
if the msx output would have scanlines, but the laserdisc frame does not).

So this patch extends PostProcessor to be able to superimpose 2 frames before
scaling. But this functionality is not yet used in this patch.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12694 7512d703-071c-0410-b5c2-d2bdd895b7d1