Fixed recording Video9000 superimpose mode
In Video9000 superimpose mode (v99x8 + v9990), both VDP outputs need to be
rendered, but only the superimposed result should end up in the video. Before
this patch both the V99x8 and the superimposed frame got recorded (and this
also caused a warning about a changing frame rate when switching between
v9990-only and v99x8+v9990-superimpose mode while recording).

You still get a waring about changing frame rate when switching between
v99x8-only and one of the two other modes, even if both VDPs are in the same
PAL/NTSC mode. The reason is that currently we don't synchronize the start of
the frames between both VDPs. So at the moment of the switch, the new frame
(from the other VDP) comes at a slightly earlier moment than expected. This
only causes a minor glitch in the video, though if there are a lot of such
switches, audio/video may get out of sync in the video.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12699 7512d703-071c-0410-b5c2-d2bdd895b7d1